添加xxhash的源码，不使用第三方库

+++ include/Read.h
@@ -23,7 +23,7 @@ namespace my_read
         Read();
         const size_t id;
         std::unique_ptr<DnaBitset> read;//序列本身每一个字符用两位表示
-        std::vector<size_t> w;//序列中间每条边的权值
+        std::vector<short> w;//序列中间每条边的权值
         size_t cnt;//余下的查询次数
     };
